requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
redis
celery
agentscope
jsonschema
hiredis
pytest
duckduckgo-search
yfinance
PyYAML
tenacity
httpx
requests
qdrant-client
ollama
python-dotenv